Reduce re-rendering of AuditTable component
