+++ components/audit_table/index.ts
@@ -5,8 +5,7 @@ import {connect} from 'react-redux';
 import {bindActionCreators, Dispatch} from 'redux';
 
 import {getMissingProfilesByIds} from 'mattermost-redux/actions/users';
-import {getUser, getCurrentUser} from 'mattermost-redux/selectors/entities/users';
-import {getChannelByName, getDirectTeammate} from 'mattermost-redux/selectors/entities/channels';
+import {getCurrentUser} from 'mattermost-redux/selectors/entities/users';
 import {GenericAction} from 'mattermost-redux/types/actions';
 
 import {GlobalState} from 'types/store';
@@ -16,9 +15,6 @@ import AuditTable from './audit_table';
 function mapStateToProps(state: GlobalState) {
     return {
         currentUser: getCurrentUser(state),
-        getUser: (userId: string) => getUser(state, userId),
-        getByName: (channelName: string) => getChannelByName(state, channelName),
-        getDirectTeammate: (channelId: string) => getDirectTeammate(state, channelId),
     };
 }
